refactor(formatter): split PmcMarkdownConverter monolith into submodules
Break the 1,151-line markdown.rs into focused submodules:
- config.rs: MarkdownConfig with grouped MetadataOptions/FigureOptions
- entities.rs: HTML entity table and clean_content()
- heading.rs: format_heading() and heading_anchor() (single source of truth)
- frontmatter.rs: YAML frontmatter generation
- metadata.rs: bold-style metadata and author formatting
- toc.rs: table of contents generation
- sections.rs: section/figure/table/additional-section conversion
- references.rs: reference formatting and conversion

Key changes:
- Unify convert/convert_with_figures (91% similar) into single method
- Unify convert_sections/convert_sections_with_figures (82% similar)
- Unify convert_figure/convert_figure_with_path duplicates
- Extract heading_anchor() as shared function for TOC and sections
- Group MarkdownConfig into MetadataOptions and FigureOptions sub-configs
- Isolate HTML entity table in entities.rs

Builder API remains fully compatible. All existing tests pass.
