+++ pubmed-cli/src/commands/markdown.rs
@@ -65,7 +65,10 @@ impl Markdown {
 
         // Prepare markdown config
         let mut config = MarkdownConfig {
-            use_yaml_frontmatter: self.frontmatter,
+            metadata: pubmed_client::pmc::MetadataOptions {
+                use_yaml_frontmatter: self.frontmatter,
+                ..Default::default()
+            },
             ..Default::default()
         };
 
@@ -115,7 +118,7 @@ impl Markdown {
                 }
             }
 
-            config.include_local_figures = true;
+            config.figures.include_local_figures = true;
             Some(figure_paths)
         } else {
             None
