+++ pubmed-formatter/src/lib.rs
@@ -19,5 +19,8 @@ pub mod pmc;
 pub mod pubmed;
 
 // Re-export main types for convenience
-pub use pmc::markdown::{HeadingStyle, MarkdownConfig, PmcMarkdownConverter, ReferenceStyle};
+pub use pmc::markdown::{
+    FigureOptions, HeadingStyle, MarkdownConfig, MetadataOptions, PmcMarkdownConverter,
+    ReferenceStyle,
+};
 pub use pubmed::export::{ExportFormat, articles_to_bibtex, articles_to_csl_json, articles_to_ris};
